finalize_confirmation(): use action as tag= param
According to all the other libraries I have seen, the tag= parameter
should be either 'allow' or 'cancel' when accepting or rejecting a
confirmation. (In other words, it should be the same as the op=
parameter.) This change does that, rather than incorrectly using
tag='conf', which I believe should only be done when retrieving
a list of confirmations.

References:
https://github.com/geel9/SteamAuth/blob/master/SteamAuth/SteamGuardAccount.cs
https://github.com/luop90/node-steam-mobile-confirmations/blob/master/index.js
https://github.com/bukson/steampy/blob/master/steampy/confirmation.py
https://github.com/Zwork101/steam-trade/blob/master/pytrade/confirmations.py
https://github.com/i-galetsky-gs-by/steamcommunity-mobile-confirmations/blob/master/index.js
https://github.com/DoctorMcKay/node-steamcommunity/blob/master/components/confirmations.js

Fixes #7